services/dspy-service/requirements.txt
# DSPy Content Quality Service
# Python 3.11+

# Core DSPy framework
dspy-ai>=2.4.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP client for external APIs
httpx>=0.26.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.5.0

# Environment management
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0  # Retry logic
structlog>=24.1.0  # Structured logging

# Optional: Local LLM support
# litellm>=1.20.0  # Unified LLM interface
